requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.2
python-dotenv==1.0.0
numpy>=1.21.0
